requirements.txt
requests
beautifulsoup4
python-dotenv
discord.py
environ-config
python-telegram-bot
